Fix circular dependency on generate_rand_color
